Update a lot of dependencies.
Updated `glam` to 0.8.1. This update removes support for the `approx`
crate so a lot of similar functionality needed to be added to
`mathbench` for tests to work. It doesn't matter for benchmarks though.

Updated `nalgebra` to 0.19.0.

Updated `euclid` to 0.20.3.

Updated `criterion` to 0.3.0 but not taking advantage of new features
yet.