Changed benchmarks to use `Bencher::iter_batched`.
This simplifies the benchmarks and I think most likely gives more
accurate results. The main differnce I think is output is stored so that
cost will be included in the benchmark, although I do need to
investigate that further.

The upshot is glam is less dominant than before, but it's probably
fairer.